`sync_protocol_error.h` moved to new path
This change is of no consequence other than correcting the inclusion
paths wherever they occur.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/53b93d7de173e2459ba70edce3401463de41a73d

commit 53b93d7de173e2459ba70edce3401463de41a73d
Author: Maksim Moskvitin <mmoskvitin@google.com>
Date:   Fri Sep 22 16:20:05 2023 +0000

    [Sync] Move sync_protocol_error.* under components/sync/engine

    While the change is motivated by crrev.com/c/4882379, it also seems
    meaningful, since SyncProtocolError seems to do a bit more than one
    would expect from protocol:util (dependency on ModelType + adding
    datatypes).

    This CL is a pure refactoring, mostly made by mass_rename.py.

    Bug: 1363089
