+++ chromium_src/components/sync/engine/sync_scheduler_impl.cc
@@ -14,7 +14,7 @@
 #include "brave/components/sync/engine/brave_sync_server_commands.h"
 
 #include "base/functional/callback.h"
-#include "components/sync/protocol/sync_protocol_error.h"
+#include "components/sync/engine/sync_protocol_error.h"
 
 namespace syncer {
