+++ components/sync/engine/brave_sync_manager_impl.h
@@ -10,7 +10,7 @@
 
 #include "base/functional/callback_forward.h"
 #include "components/sync/engine/sync_manager_impl.h"
-#include "components/sync/protocol/sync_protocol_error.h"
+#include "components/sync/engine/sync_protocol_error.h"
 
 namespace syncer {
